use a generic push down button instead of duplicating logic
